Refactor code for routes configuration.
