+++ auth-server/src/main/java/io/microservices/auth/server/AuthServerApplication.java
@@ -1,4 +1,4 @@
-package io.microservices.apigateway;
+package io.microservices.auth.server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,10 +6,10 @@ import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 
 @EnableEurekaClient
 @SpringBootApplication
-public class ApiGatewayApplication {
+public class AuthServerApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(ApiGatewayApplication.class, args);
+        SpringApplication.run(AuthServerApplication.class, args);
     }
 
 }
