+++ auth-server/src/main/resources/application-prod.yml
@@ -1,0 +1,4 @@
+management:
+  endpoint:
+    health:
+      show-details: when_authorized
