--- authorization-server/src/main/java/io/microservices/authorization/server/AuthorizationServerApplication.java
@@ -1,13 +1,0 @@
-package io.microservices.authorization.server;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AuthorizationServerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AuthorizationServerApplication.class, args);
-    }
-
-}
